wifi: (refs #414) Do not truncate Thompson Sampling Beta parameters
ThompsonSamplingWifiManager::SampleBetaVariable took its alpha/beta
parameters as uint64_t, but it is called with 1.0 + the EWMA-averaged
success/failure counts, which are fractional doubles. The fractional part
was therefore truncated, discarding the small success/failure differences
between competing rates. This makes Thompson sampling converge slowly to
(or get stuck on) a poor rate for some random seeds, which is the
behaviour reported in issue #414 ("0 throughput when changing the random
seed"). Take the parameters as double so the real-valued Beta parameters
are preserved.

No existing test exercises ThompsonSampling rate selection, so this is a
behavioural improvement without a deterministic automated regression (the
reported symptom is statistical across seeds); flagged refs (not fixes)
for maintainer review.

(cherry picked from commit 9f1284b7e8faf54e148417c83e8cd7a7b5cb5e04)

+++ src/wifi/model/rate-control/thompson-sampling-wifi-manager.cc
@@ -404,8 +404,13 @@ ThompsonSamplingWifiManager::DoGetRtsTxVector(WifiRemoteStation* st)
 }
 
 double
-ThompsonSamplingWifiManager::SampleBetaVariable(uint64_t alpha, uint64_t beta) const
+ThompsonSamplingWifiManager::SampleBetaVariable(double alpha, double beta) const
 {
+    // alpha and beta are the real-valued Beta parameters (1 + the EWMA-averaged
+    // success/failure counts). They must not be truncated to integers, otherwise
+    // small differences in the success/failure estimates of competing rates are
+    // lost and Thompson sampling can converge slowly to (or get stuck on) a poor
+    // rate for some random seeds (issue #414).
     double X = m_gammaRandomVariable->GetValue(alpha, 1.0);
     double Y = m_gammaRandomVariable->GetValue(beta, 1.0);
     return X / (X + Y);
